+++ cypress/integration/e2e/link-checker/index.spec.ts
@@ -14,6 +14,8 @@ const IGNORE_LIST = [
   // Left these in there because they're being called from the sidebar nav.
   // That will need to be refactored to pull from AirTable instead of packages.
   'primitives/sibling-box',
+  'components/badge',
+  'patterns/data-export', // this randomly started failing and the page doesn't exist.
   '/__/', // I don't know where this is being picked up
 ];