+++ packages/paste-core/components/list/stories/index.stories.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
+import {CustomizationProvider} from '@twilio-paste/customization';
 import {Paragraph} from '@twilio-paste/paragraph';
+import {Stack} from '@twilio-paste/stack';
 import {OrderedList, UnorderedList, ListItem} from '../src';
 
 // eslint-disable-next-line import/no-default-export
@@ -14,28 +16,27 @@ export const DefaultOrderedList = (): React.ReactNode => {
     <>
       <OrderedList data-test="something" aria-label="ordered list">
         <ListItem>
-          Kale chips distillery authentic, portland etsy cloud bread vinyl gentrify drinking vinegar viral meh hot
-          chicken bitters fashion axe palo santo. Chillwave fixie sustainable <i>helvetica</i> etsy.
+          Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+          API.
         </ListItem>
         <ListItem>
-          Prism whatever ethical, gochujang <strong>edison bulb</strong> put a bird on it kitsch. Pop-up 90&apos;s la
-          croix tumeric, palo santo chia try-hard direct trade tote bag roof party scenester kitsch stumptown
-          intelligentsia. Literally heirloom blue bottle etsy.
+          Proactively inform customers about account activity, purchase confirmations, and shipping notifications with
+          the
+          <i>Programmable Messaging API</i>.
         </ListItem>
       </OrderedList>
       <Paragraph>
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Sequi nemo doloremque qui a vitae officiis laborum
-        dolor. Pariatur obcaecati sequi minus nobis eos mollitia velit dolor, quis amet accusamus officiis?
+        Build customer relationships with secure, cross-channel conversations that can include multiple parties, all on
+        one platform with the Conversations API.
       </Paragraph>
       <OrderedList aria-label="styled">
         <ListItem>
-          Kale chips distillery authentic, portland etsy cloud bread vinyl gentrify drinking vinegar viral meh hot
-          chicken bitters fashion axe palo santo. Chillwave fixie sustainable <i>helvetica</i> etsy.
+          Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+          API.
         </ListItem>
         <ListItem>
-          Prism whatever ethical, gochujang <strong>edison bulb</strong> put a bird on it kitsch. Pop-up 90&apos;s la
-          croix tumeric, palo santo chia try-hard direct trade tote bag roof party scenester kitsch stumptown
-          intelligentsia. Literally heirloom blue bottle etsy.
+          Proactively inform customers about <strong>account activity</strong>, purchase confirmations, and shipping
+          notifications with the Programmable Messaging API.
         </ListItem>
       </OrderedList>
     </>
@@ -51,24 +52,23 @@ export const DefaultUnorderedList = (): React.ReactNode => {
     <>
       <UnorderedList>
         <ListItem data-test="test-hook" title="some title">
-          Kale chips distillery authentic, portland etsy cloud bread vinyl gentrify drinking vinegar viral meh hot
-          chicken bitters fashion axe palo santo. Chillwave fixie sustainable <i>helvetica</i> etsy.
+          Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+          API.
         </ListItem>
         <ListItem>
-          Prism whatever ethical, gochujang <strong>edison bulb</strong> put a bird on it kitsch. Pop-up 90&apos;s la
-          croix tumeric, palo santo chia try-hard direct trade tote bag roof party scenester kitsch stumptown
-          intelligentsia. Literally heirloom blue bottle etsy.
+          Proactively inform customers about account activity, purchase confirmations, and shipping notifications with
+          the
+          <i>Programmable Messaging API</i>.
         </ListItem>
       </UnorderedList>
       <UnorderedList>
         <ListItem>
-          Kale chips distillery authentic, portland etsy cloud bread vinyl gentrify drinking vinegar viral meh hot
-          chicken bitters fashion axe palo santo. Chillwave fixie sustainable <i>helvetica</i> etsy.
+          Build customer relationships with secure, cross-channel conversations that can include multiple parties, all
+          on one platform with the Conversations API.
         </ListItem>
         <ListItem>
-          Prism whatever ethical, gochujang <strong>edison bulb</strong> put a bird on it kitsch. Pop-up 90&apos;s la
-          croix tumeric, palo santo chia try-hard direct trade tote bag roof party scenester kitsch stumptown
-          intelligentsia. Literally heirloom blue bottle etsy.
+          Proactively inform customers about <strong>account activity</strong>, purchase confirmations, and shipping
+          notifications with the Programmable Messaging API.
         </ListItem>
       </UnorderedList>
     </>
@@ -84,30 +84,175 @@ export const NestedUnorderedList = (): React.ReactNode => {
     <>
       <UnorderedList>
         <ListItem data-test="test-hook-nested" title="nested list title">
-          Kale chips distillery authentic, portland etsy cloud bread vinyl gentrify drinking vinegar viral meh hot
-          chicken bitters fashion axe palo santo. Chillwave fixie sustainable <i>helvetica</i> etsy.
+          Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+          API.
           <UnorderedList marginBottom="space0">
             <ListItem>
-              Kale chips distillery authentic, portland etsy cloud bread vinyl gentrify drinking vinegar viral meh hot
-              chicken bitters fashion axe palo santo. Chillwave fixie sustainable <i>helvetica</i> etsy.
+              Proactively inform customers about account activity, purchase confirmations, and shipping notifications
+              with the
+              <i>Programmable Messaging API</i>.
             </ListItem>
             <ListItem>
-              Prism whatever ethical, gochujang <strong>edison bulb</strong> put a bird on it kitsch. Pop-up 90&apos;s
-              la croix tumeric, palo santo chia try-hard direct trade tote bag roof party scenester kitsch stumptown
-              intelligentsia. Literally heirloom blue bottle etsy.
+              Build customer relationships with secure, cross-channel conversations that can include multiple parties,
+              all on one platform with the Conversations API.
             </ListItem>
           </UnorderedList>
         </ListItem>
         <ListItem>
-          Prism whatever ethical, gochujang <strong>edison bulb</strong> put a bird on it kitsch. Pop-up 90&apos;s la
-          croix tumeric, palo santo chia try-hard direct trade tote bag roof party scenester kitsch stumptown
-          intelligentsia. Literally heirloom blue bottle etsy.
+          Proactively inform customers about <strong>account activity</strong>, purchase confirmations, and shipping
+          notifications with the Programmable Messaging API.
         </ListItem>
         <ListItem>
-          Kale chips distillery authentic, portland etsy cloud bread vinyl gentrify drinking vinegar viral meh hot
-          chicken bitters fashion axe palo santo. Chillwave fixie sustainable <i>helvetica</i> etsy.
+          Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+          API.
         </ListItem>
       </UnorderedList>
     </>
+  );
+};
+
+export const CustomOrderedList: React.FC = () => {
+  return (
+    <Stack orientation="vertical" spacing="space60">
+      <OrderedList aria-label="ordered list">
+        <ListItem>
+          Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+          API.
+        </ListItem>
+        <ListItem>
+          Proactively inform customers about account activity, purchase confirmations, and shipping notifications with
+          the
+          <i>Programmable Messaging API</i>.
+        </ListItem>
+      </OrderedList>
+      <CustomizationProvider
+        baseTheme="default"
+        elements={{
+          ORDERED_LIST: {
+            backgroundColor: 'colorBackgroundPrimaryWeaker',
+            marginLeft: 'space0',
+            padding: 'space40',
+            paddingLeft: 'space70',
+          },
+          LIST_ITEM: {
+            color: 'colorTextError',
+          },
+        }}
+      >
+        <OrderedList aria-label="ordered list">
+          <ListItem>
+            Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+            API.
+          </ListItem>
+          <ListItem>
+            Proactively inform customers about account activity, purchase confirmations, and shipping notifications with
+            the
+            <i>Programmable Messaging API</i>.
+          </ListItem>
+        </OrderedList>
+      </CustomizationProvider>
+      <CustomizationProvider
+        baseTheme="default"
+        elements={{
+          NEW_ORDERED_LIST: {
+            backgroundColor: 'colorBackgroundTrial',
+            marginLeft: 'space0',
+            padding: 'space40',
+            paddingLeft: 'space70',
+          },
+          NEW_LIST_ITEM: {
+            color: 'colorTextLink',
+          },
+          ANOTHER_NEW_LIST_ITEM: {
+            color: 'colorTextErrorStrong',
+          },
+        }}
+      >
+        <OrderedList aria-label="ordered list" element="NEW_ORDERED_LIST">
+          <ListItem element="NEW_LIST_ITEM">
+            Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+            API.
+          </ListItem>
+          <ListItem element="ANOTHER_NEW_LIST_ITEM">
+            Proactively inform customers about account activity, purchase confirmations, and shipping notifications with
+            the
+            <i>Programmable Messaging API</i>.
+          </ListItem>
+        </OrderedList>
+      </CustomizationProvider>
+    </Stack>
+  );
+};
+
+export const CustomUnorderedList: React.FC = () => {
+  return (
+    <Stack orientation="vertical" spacing="space60">
+      <UnorderedList aria-label="ordered list">
+        <ListItem>
+          Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+          API.
+        </ListItem>
+        <ListItem>
+          Proactively inform customers about account activity, purchase confirmations, and shipping notifications with
+          the
+          <i>Programmable Messaging API</i>.
+        </ListItem>
+      </UnorderedList>
+      <CustomizationProvider
+        baseTheme="default"
+        elements={{
+          UNORDERED_LIST: {
+            backgroundColor: 'colorBackgroundPrimaryWeaker',
+            marginLeft: 'space0',
+            padding: 'space40',
+            paddingLeft: 'space70',
+          },
+          LIST_ITEM: {
+            color: 'colorTextError',
+          },
+        }}
+      >
+        <UnorderedList aria-label="ordered list">
+          <ListItem>
+            Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+            API.
+          </ListItem>
+          <ListItem>
+            Proactively inform customers about account activity, purchase confirmations, and shipping notifications with
+            the
+            <i>Programmable Messaging API</i>.
+          </ListItem>
+        </UnorderedList>
+      </CustomizationProvider>
+      <CustomizationProvider
+        baseTheme="default"
+        elements={{
+          NEW_UNORDERED_LIST: {
+            backgroundColor: 'colorBackgroundTrial',
+            marginLeft: 'space0',
+            padding: 'space40',
+            paddingLeft: 'space70',
+          },
+          NEW_LIST_ITEM: {
+            color: 'colorTextLink',
+          },
+          ANOTHER_NEW_LIST_ITEM: {
+            color: 'colorTextErrorStrong',
+          },
+        }}
+      >
+        <UnorderedList aria-label="ordered list" element="NEW_UNORDERED_LIST">
+          <ListItem element="NEW_LIST_ITEM">
+            Deliver critical time-sensitive messages to employees and customers at scale with the Programmable Messaging
+            API.
+          </ListItem>
+          <ListItem element="ANOTHER_NEW_LIST_ITEM">
+            Proactively inform customers about account activity, purchase confirmations, and shipping notifications with
+            the
+            <i>Programmable Messaging API</i>.
+          </ListItem>
+        </UnorderedList>
+      </CustomizationProvider>
+    </Stack>
   );
 };