+++ packages/paste-style-props/src/proptypes/space.ts
@@ -8,3 +8,4 @@ const MarginOptions = ['auto', ...Object.keys(DefaultTheme.space)];
 export const isSpaceTokenProp = propValidator(SpaceOptions);
 export const isPaddingTokenProp = propValidator(SpaceOptions);
 export const isMarginTokenProp = propValidator(MarginOptions);
+export const isGapTokenProp = propValidator(SpaceOptions);